KG_RAG_backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
neo4j
pypdf2==3.0.1
spacy==3.7.4
sentence-transformers==2.5.1
langchain
openai==1.12.0
scikit-learn==1.4.2
langchain_openai
langchain-core
langchain-neo4j
langchain-community
langchain-huggingface